requirements.txt
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
humanfriendly==10.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
onnx==1.20.1
onnxruntime==1.23.2
packaging==26.0
pandas==3.0.0
pillow==12.1.0
protobuf==6.33.5
psutil==7.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scikit-learn==1.8.0
scipy==1.17.0
setuptools==70.2.0
six==1.17.0
skl2onnx==1.20.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0+cpu
torchvision==0.25.0+cpu
tqdm==4.67.3
typing_extensions==4.15.0
